Add parser and fix bug
