+++ app/src/main/java/com/kantai_wiki/kcw_mobile/util/HTTPUtils.java
@@ -1,30 +1,15 @@
 package com.kantai_wiki.kcw_mobile.util;
 
 import android.content.Context;
-import android.content.pm.LabeledIntent;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
-
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
-import com.kantai_wiki.kcw_mobile.model.EquipmentEnemy;
-import com.kantai_wiki.kcw_mobile.model.EquipmentKMS;
-import com.kantai_wiki.kcw_mobile.model.EquipmentUpgrade;
-import com.kantai_wiki.kcw_mobile.model.Expedition;
-import com.kantai_wiki.kcw_mobile.model.Quest;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.io.Reader;
-import java.io.UnsupportedEncodingException;
-import java.lang.reflect.Type;
 import java.net.HttpURLConnection;
-import java.net.URI;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * Created by Wafer on 2015/8/30.
@@ -33,7 +18,7 @@ import java.util.List;
 public class HTTPUtils {
 
     //checkConnection
-    public static boolean checkConnection(){
+    public static boolean checkConnection() {
 
         boolean flag;
 
@@ -48,30 +33,38 @@ public class HTTPUtils {
 
 
     //download
-    public static String download(String address) throws IOException{
+    public static String download(String address) throws IOException {
         InputStream inputStream = null;
 
         int READ_TIME_OUT_LIMIT = 10000;
         int CONNECT_TIME_OUT_LIMIT = 15000;
 
+        HttpURLConnection connection = null;
         try {
             URL url = new URL(address);
-            HttpURLConnection connection = (HttpURLConnection) url.openConnection();
+            connection = (HttpURLConnection) url.openConnection();
 
             //set read and connect time out
             connection.setReadTimeout(READ_TIME_OUT_LIMIT);
             connection.setConnectTimeout(CONNECT_TIME_OUT_LIMIT);
-
-            //set request method
             connection.setRequestMethod("GET");
+
+            //start query
+            connection.connect();
 
             //get input stream
             inputStream = connection.getInputStream();
             return readInputStream(inputStream);
-        }
-        finally {
+        } finally {
+
+            //close inputStream
             if (inputStream != null) {
                 inputStream.close();
+            }
+
+            //close connection
+            if (connection != null) {
+                connection.disconnect();
             }
         }
     }
@@ -90,49 +83,4 @@ public class HTTPUtils {
         return stringBuilder.toString();
     }
 
-    /**
-     * The parser with GSON
-     *
-     */
-
-    //Quest
-    public List<Quest> praseQuest (String jsonString) {
-
-        Gson gson = new Gson();
-        Type listType = new TypeToken<List<Quest>>(){}.getType();
-
-        return gson.fromJson(jsonString, listType);
-    }
-
-    //Expedition
-    public List<Expedition> praseExpedition (String jsonString) {
-        Gson gson = new Gson();
-        Type listType = new  TypeToken<List<Expedition>>(){}.getType();
-
-        return gson.fromJson(jsonString,listType);
-    }
-
-    //EquipmentKMS
-    public List<EquipmentKMS> parseEquipmentKMS (String jsonString) {
-        Gson gson = new Gson();
-        Type listType = new TypeToken<List<EquipmentKMS>>() {}.getType();
-
-        return gson.fromJson(jsonString,listType);
-    }
-
-    //EquipmentEnemy
-    public List<EquipmentEnemy> parseEquipmentEnemy (String jsonString) {
-        Gson gson = new Gson();
-        Type listType = new TypeToken<List<EquipmentEnemy>>() {}.getType();
-
-        return gson.fromJson(jsonString,listType);
-    }
-
-    //EquipmentUpgrade
-    public List<EquipmentUpgrade> parseEquipmentUpgrade (String jsonString) {
-        Gson gson = new Gson();
-        Type listType = new TypeToken<List<EquipmentUpgrade>>() {}.getType();
-
-        return gson.fromJson(jsonString,listType);
-    }
 }
